large docs refactoring
