+++ examples/plug-and-play/demo_PnP_DPIR_deblur.py
@@ -19,7 +19,7 @@ from deepinv.optim.prior import PnP
 from deepinv.optim.optimizers import optim_builder
 from deepinv.training_utils import test
 from torchvision import transforms
-from deepinv.utils.parameters import get_DPIR_params
+from deepinv.optim.dpir import get_DPIR_params
 from deepinv.utils.demo import load_dataset, load_degradation
 
 # %%
@@ -106,6 +106,9 @@ dataset = dinv.datasets.HDF5Dataset(path=dinv_dataset_path, train=True)
 # This method is based on half-quadratic splitting (HQS).
 # The algorithm alternates between a denoising step and a data fidelity step, where
 # the denoising step is performed by a pretrained denoiser :class:`deepinv.models.DRUNet`.
+#
+# .. note::
+#    We provide a wrapper for rapidly creating the DPIR algorithm in :class:`deepinv.optim.DPIR`.
 
 # load specific parameters for DPIR
 lamb, sigma_denoiser, stepsize, max_iter = get_DPIR_params(noise_level_img)
@@ -133,6 +136,7 @@ model = optim_builder(
 # Evaluate the model on the problem.
 # --------------------------------------------------------------------
 # The test function evaluates the model on the test dataset and computes the metrics.
+#
 
 save_folder = RESULTS_DIR / method / operation / dataset_name
 wandb_vis = False  # plot curves and images in Weight&Bias.
